pyproject.toml
[tool.poetry]
name = "mediaconverter"
version = "0.1.0"
description = ""
authors = ["Xie Yanbo"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.14.0"
Django = "^4.1.7"
django-filter = "^22.1"
python-ffmpeg = "^2.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
pytest-coverage = "^0.0"
pytest-django = "^4.5.2"
pre-commit = "^3.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
